Use hakari to keep the project compiling more quickly
With the various features we select for crates, every time
rust-analyzer runs in a certain way, or you run a binary and then
another binary, or you sneeze, cargo rebuilds everything afresh.

Better to put all the deps in the workspace-hack (a pattern pioneered
by mozilla), managed by cargo-hakari. I don't love this but it seems
to work?