Make list_bookmarks_with_tag into list_bookmarks_matching with criteria
Using a trait, we can pretty easily make a basically unified
bookmark-finding operation that constructs the right SQL query based
on the search criteria it's given.

It could use some documentation, but in my experiments, this has
worked pretty great so far. I think we can make this the single
bookmark-querying function (if it starts doing the right thing when
given no criteria). That'd be pretty awesome.