Link associations
This changeset adds a new `primary_link` column in thie bookmarks, and
a new `bookmark_associations` table, for capturing a distinction between
normal, primary links--those saved in their own right--and those
representing supplementary information (e.g. a Lobsters thread or
MetaFilter discussion) which are valuable mainly in the context of the
primary link. An assocation can currently be added only via the optional
`associated-link` (and `associated-context`) flags to the `add` CLI
command.