Define all tables as STRICT
That ensures that strings can't be stored in integer columns (??!?!)

But: it also means that the things we *thought* were types (JSON,
DATETIME, BOOLEAN) - aren't. We have to specify them as the types that
sqlite stores them as, and then do more type juggling. Thankfully sqlx
is strict, but ugh.