Split read-write and read-only operations into two separate DB pools
The rw pool allows only one connection, which in combination with WAL
mode means that we should never see the dreaded "sqlite busy" error,
while still being able to fire off as many readonly transactions as we
want.

This makes use of a pretty neat type-level trick, where methods that
apply to the read-only transaction are defined on both the RO and the
RW txn type, but write operations can only be accessed on (because
they're only defined on) the read-write transaction.