Data model: Extract all urls into separate table `urls`, associate those
I am not super happy with the primary_link flag on bookmarks; it is
kinda a denormalized "is this a real bookmark" flag, which is a great
indicator that something should be broken out! This change does that:
It takes the `url` field on the `bookmarks` table and moves it to a
separate table, in `urls.links`.

This means we can do the exact same thing as before, just with a bit
more finesse: associations start in the `urls` table, and if somebody
wants to upgrade them to a full bookmark, that's easy - and it means
our bookmark-handling code can be blissfully unaware of associated
links that aren't full bookmarks yet... because they're not bookmarks,
with or without the flag.

This is a bit of a larger refactor, but it does seem to do the trick:
Tests pass, I can import bookmarks and the web app does the thing. It
also reduced quite a bit of logic in the CLI app's add_cmd
function.