Fix listing bookmarks such that they update ... once.
I'm losing my mind with dioxus, and I think this rewrite is basically
non-salvageable. That it's this difficult to tell what is going on is
a strong indicator to me that it's not ready for serious (or even
hobbyist, by-me) usage, at least now; but maybe never will be.

Maybe reactive programming just isn't for me?! Who knows.