Rework the tag selector to support editing & deletion
Instead of doing a single multi-select box, do a scheme that keeps
track of tags we want (selected) and a current tag that's being
selected (active); uses a state machine 2 layers deep. It's not
perfect yet - requires a click/tab&enter to enter another tag... but
it does at least not get stuck.